+++ crates/api/src/cfg/file.rs
@@ -1440,35 +1440,6 @@ pub struct InitialObjectsConfig {
     pub networks: Option<HashMap<String, NetworkDefinition>>,
 }
 
-impl DpaConfig {
-    pub const fn default_hb_interval() -> chrono::TimeDelta {
-        chrono::TimeDelta::minutes(2)
-    }
-
-    pub const fn default_monitor_run_interval() -> std::time::Duration {
-        std::time::Duration::from_secs(60)
-    }
-
-    pub const fn default_subnet_ip() -> Ipv4Addr {
-        Ipv4Addr::UNSPECIFIED
-    }
-}
-
-impl Default for DpaConfig {
-    fn default() -> Self {
-        Self {
-            enabled: false,
-            mqtt_endpoint: default_mqtt_endpoint(),
-            mqtt_broker_port: default_mqtt_broker_port(),
-            subnet_ip: Self::default_subnet_ip(),
-            subnet_mask: 0,
-            hb_interval: Self::default_hb_interval(),
-            monitor_run_interval: Self::default_monitor_run_interval(),
-            auth: MqttAuthConfig::default(),
-        }
-    }
-}
-
 /// TLS certificate and key configuration for securing
 /// gRPC and HTTP connections.
 #[derive(Clone, Debug, Deserialize, Serialize)]
@@ -2018,116 +1989,7 @@ fn default_mqtt_broker_port() -> u16 {
     1884
 }
 
-/// MQTT authentication mode.
-#[derive(Clone, Debug, Default, Deserialize, Serialize, PartialEq)]
-#[serde(rename_all = "snake_case")]
-pub enum MqttAuthMode {
-    /// No authentication.
-    #[default]
-    None,
-    /// Username/password basic authentication.
-    BasicAuth,
-    /// OAuth2 token-based authentication.
-    Oauth2,
-}
-
-/// OAuth2 configuration for MQTT broker authentication.
-#[derive(Clone, Debug, Deserialize, Serialize, PartialEq)]
-pub struct MqttOAuth2Config {
-    /// OAuth2 token endpoint URL.
-    pub token_url: String,
-
-    /// OAuth2 scopes to request when obtaining a token.
-    #[serde(default)]
-    pub scopes: Vec<String>,
-
-    /// HTTP timeout for token endpoint requests.
-    /// Default is 30 seconds.
-    #[serde(
-        default = "MqttOAuth2Config::default_http_timeout",
-        deserialize_with = "deserialize_duration",
-        serialize_with = "as_std_duration"
-    )]
-    pub http_timeout: std::time::Duration,
-
-    /// Username sent with the MQTT CONNECT packet when using
-    /// OAuth2.
-    /// Default is "oauth2token".
-    #[serde(default = "MqttOAuth2Config::default_username")]
-    pub username: String,
-}
-
-impl MqttOAuth2Config {
-    fn default_http_timeout() -> std::time::Duration {
-        std::time::Duration::from_secs(30)
-    }
-
-    fn default_username() -> String {
-        "oauth2token".to_string()
-    }
-}
-
-/// MQTT authentication configuration shared by DPA and
-/// DSX event bus.
-#[derive(Clone, Debug, Default, Deserialize, Serialize, PartialEq)]
-pub struct MqttAuthConfig {
-    /// Authentication mechanism to use for MQTT
-    /// connections.
-    #[serde(default)]
-    pub auth_mode: MqttAuthMode,
-
-    /// OAuth2 settings, required when `auth_mode` is
-    /// `Oauth2`.
-    pub oauth2: Option<MqttOAuth2Config>,
-}
-
-/// DPA (aka Cluster Interconnect Network) related configuration
-/// Enabled DPA, and specifies basic network settings.
-/// The VNI to be used by DPA will be the same as the parent VPC.
-#[derive(Clone, Debug, Deserialize, Serialize, PartialEq)]
-pub struct DpaConfig {
-    /// Global enable/disable of Cluster Interconnect Network
-    #[serde(default)]
-    pub enabled: bool,
-
-    /// MQTT broker host (name or IP address) used to create client connections
-    #[serde(default = "default_mqtt_endpoint")]
-    pub mqtt_endpoint: String,
-
-    /// MQTT broker port to use to estabilsh client connections
-    #[serde(default = "default_mqtt_broker_port")]
-    pub mqtt_broker_port: u16,
-
-    /// Base IPv4 address of the DPA/Cluster Interconnect
-    /// subnet.
-    #[serde(default = "DpaConfig::default_subnet_ip")]
-    pub subnet_ip: Ipv4Addr,
-
-    /// CIDR prefix length for the DPA subnet.
-    #[serde(default)]
-    pub subnet_mask: i32,
-
-    /// hb_interval is the interval at which we issue heartbeat
-    /// requests to the DPA.
-    /// Defaults to 120 if not specified.
-    #[serde(
-        default = "DpaConfig::default_hb_interval",
-        deserialize_with = "deserialize_duration_chrono",
-        serialize_with = "as_duration"
-    )]
-    pub hb_interval: chrono::TimeDelta,
-
-    /// The interval at which we run the DPA monitor.
-    #[serde(
-        default = "DpaConfig::default_monitor_run_interval",
-        deserialize_with = "deserialize_duration",
-        serialize_with = "as_std_duration"
-    )]
-    pub monitor_run_interval: std::time::Duration,
-
-    #[serde(default)]
-    pub auth: MqttAuthConfig,
-}
+pub use carbide_dpa_manager::cfg::{DpaConfig, MqttAuthConfig, MqttAuthMode};
 
 /// DSX Exchange Event Bus configuration for publishing state change events via MQTT 3.1.1.
 ///
